feat: add await termination for scheduler
