+++ src/storage/src/compaction/scheduler.rs
@@ -109,240 +109,3 @@ where
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::time::Duration;
-
-    use super::*;
-    use crate::scheduler::dedup_deque::DedupDeque;
-    use crate::scheduler::rate_limit::{
-        BoxedRateLimitToken, CascadeRateLimiter, MaxInflightTaskLimiter,
-    };
-    use crate::scheduler::{HandlerLoop, LocalScheduler, Scheduler, SchedulerConfig};
-
-    struct CountdownLatch {
-        counter: std::sync::Mutex<usize>,
-        notifies: std::sync::RwLock<Vec<Arc<Notify>>>,
-    }
-
-    impl CountdownLatch {
-        fn new(size: usize) -> Self {
-            Self {
-                counter: std::sync::Mutex::new(size),
-                notifies: std::sync::RwLock::new(vec![]),
-            }
-        }
-
-        fn countdown(&self) {
-            let mut counter = self.counter.lock().unwrap();
-            if *counter >= 1 {
-                *counter -= 1;
-                if *counter == 0 {
-                    let notifies = self.notifies.read().unwrap();
-                    for waiter in notifies.iter() {
-                        waiter.notify_one();
-                    }
-                }
-            }
-        }
-
-        async fn wait(&self) {
-            let notify = Arc::new(Notify::new());
-            {
-                let notify = notify.clone();
-                let mut notifies = self.notifies.write().unwrap();
-                notifies.push(notify);
-            }
-            notify.notified().await
-        }
-    }
-
-    #[tokio::test]
-    async fn test_schedule_handler() {
-        common_telemetry::init_default_ut_logging();
-        let queue = Arc::new(std::sync::RwLock::new(DedupDeque::default()));
-        let latch = Arc::new(CountdownLatch::new(2));
-        let latch_cloned = latch.clone();
-        let handler = Arc::new(HandlerLoop {
-            req_queue: queue.clone(),
-            cancel_token: Default::default(),
-            task_notifier: Arc::new(Default::default()),
-            request_handler: MockHandler {
-                cb: move || {
-                    latch_cloned.countdown();
-                },
-            },
-            limiter: Arc::new(CascadeRateLimiter::new(vec![Box::new(
-                MaxInflightTaskLimiter::new(3),
-            )])),
-        });
-
-        let handler_cloned = handler.clone();
-        common_runtime::spawn_bg(async move { handler_cloned.run().await });
-
-        queue.write().unwrap().push_back(1, MockRequest::default());
-        handler.task_notifier.notify_one();
-        queue.write().unwrap().push_back(2, MockRequest::default());
-        handler.task_notifier.notify_one();
-
-        tokio::time::timeout(Duration::from_secs(1), latch.wait())
-            .await
-            .unwrap();
-    }
-
-    #[derive(Default, Debug)]
-    struct MockRequest {
-        region_id: RegionId,
-    }
-
-    struct MockHandler<F> {
-        cb: F,
-    }
-
-    #[async_trait::async_trait]
-    impl<F> Handler for MockHandler<F>
-    where
-        F: Fn() + Send + Sync,
-    {
-        type Request = MockRequest;
-
-        async fn handle_request(
-            &self,
-            _req: Self::Request,
-            token: BoxedRateLimitToken,
-            finish_notifier: Arc<Notify>,
-        ) -> Result<()> {
-            (self.cb)();
-            token.try_release();
-            finish_notifier.notify_one();
-            Ok(())
-        }
-    }
-
-    impl Request for MockRequest {
-        type Key = RegionId;
-
-        fn key(&self) -> Self::Key {
-            self.region_id
-        }
-    }
-
-    #[tokio::test]
-    async fn test_scheduler() {
-        let latch = Arc::new(CountdownLatch::new(2));
-        let latch_cloned = latch.clone();
-
-        let handler = MockHandler {
-            cb: move || {
-                latch_cloned.countdown();
-            },
-        };
-        let scheduler: LocalScheduler<MockRequest> = LocalScheduler::new(
-            SchedulerConfig {
-                max_inflight_task: 3,
-            },
-            handler,
-        );
-
-        scheduler.schedule(MockRequest { region_id: 1 }).unwrap();
-
-        scheduler.schedule(MockRequest { region_id: 2 }).unwrap();
-
-        tokio::time::timeout(Duration::from_secs(1), latch.wait())
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_scheduler_many() {
-        common_telemetry::init_default_ut_logging();
-        let task_size = 100;
-
-        let latch = Arc::new(CountdownLatch::new(task_size));
-        let latch_clone = latch.clone();
-
-        let handler = MockHandler {
-            cb: move || {
-                latch_clone.countdown();
-            },
-        };
-
-        let config = SchedulerConfig {
-            max_inflight_task: 3,
-        };
-        let scheduler = LocalScheduler::new(config, handler);
-
-        for i in 0..task_size {
-            scheduler
-                .schedule(MockRequest {
-                    region_id: i as RegionId,
-                })
-                .unwrap();
-        }
-
-        tokio::time::timeout(Duration::from_secs(3), latch.wait())
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_scheduler_interval() {
-        common_telemetry::init_default_ut_logging();
-        let task_size = 100;
-        let latch = Arc::new(CountdownLatch::new(task_size));
-        let latch_clone = latch.clone();
-
-        let handler = MockHandler {
-            cb: move || {
-                latch_clone.countdown();
-            },
-        };
-
-        let config = SchedulerConfig {
-            max_inflight_task: 3,
-        };
-        let scheduler = LocalScheduler::new(config, handler);
-
-        for i in 0..task_size / 2 {
-            scheduler
-                .schedule(MockRequest {
-                    region_id: i as RegionId,
-                })
-                .unwrap();
-        }
-
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        for i in task_size / 2..task_size {
-            scheduler
-                .schedule(MockRequest {
-                    region_id: i as RegionId,
-                })
-                .unwrap();
-        }
-
-        tokio::time::timeout(Duration::from_secs(6), latch.wait())
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_schedule_duplicate_tasks() {
-        common_telemetry::init_default_ut_logging();
-        let handler = MockHandler { cb: || {} };
-        let config = SchedulerConfig {
-            max_inflight_task: 3,
-        };
-        let scheduler = LocalScheduler::new(config, handler);
-
-        let mut scheduled_task = 0;
-        for _ in 0..10 {
-            if scheduler.schedule(MockRequest { region_id: 1 }).unwrap() {
-                scheduled_task += 1;
-            }
-        }
-        scheduler.stop().await.unwrap();
-        debug!("Schedule tasks: {}", scheduled_task);
-        assert!(scheduled_task < 10);
-    }
-}
