+++ src/storage/src/scheduler.rs
@@ -14,17 +14,18 @@
 
 use std::fmt::{Debug, Formatter};
 use std::hash::Hash;
+use std::sync::atomic::{AtomicI32, Ordering};
 use std::sync::{Arc, Mutex, RwLock};
 
 use async_trait::async_trait;
 use common_telemetry::{debug, error, info};
-use snafu::ResultExt;
+use snafu::{ensure, ResultExt};
 use tokio::sync::Notify;
 use tokio::task::JoinHandle;
 use tokio_util::sync::CancellationToken;
 
 use crate::error;
-use crate::error::StopSchedulerSnafu;
+use crate::error::{IllegalSchedulerStateSnafu, StopSchedulerSnafu};
 use crate::scheduler::dedup_deque::DedupDeque;
 use crate::scheduler::rate_limit::{
     BoxedRateLimitToken, CascadeRateLimiter, MaxInflightTaskLimiter, RateLimiter,
@@ -64,8 +65,9 @@ pub trait Scheduler: Debug {
     /// contains the request with same key.
     fn schedule(&self, request: Self::Request) -> error::Result<bool>;
 
-    /// Stops scheduler.
-    async fn stop(&self) -> error::Result<()>;
+    /// Stops scheduler. If `await_termination` is set to true, the scheduler will
+    /// wait until all queued requests are processed.
+    async fn stop(&self, await_termination: bool) -> error::Result<()>;
 }
 
 /// Scheduler config.
@@ -82,6 +84,10 @@ impl Default for SchedulerConfig {
     }
 }
 
+const STATE_RUNNING: i32 = 0;
+const STATE_STOP: i32 = 1;
+const STATE_AWAIT_TERMINATION: i32 = 2;
+
 /// Request scheduler based on local state.
 pub struct LocalScheduler<R: Request> {
     /// Request FIFO with key deduplication.
@@ -92,6 +98,8 @@ pub struct LocalScheduler<R: Request> {
     task_notifier: Arc<Notify>,
     /// Join handle of spawned request handling loop.
     join_handle: Mutex<Option<JoinHandle<()>>>,
+    /// State of scheduler.
+    state: Arc<AtomicI32>,
 }
 
 impl<R> Debug for LocalScheduler<R>
@@ -111,6 +119,7 @@ where
     type Request = R;
 
     fn schedule(&self, request: Self::Request) -> error::Result<bool> {
+        ensure!(self.running(), IllegalSchedulerStateSnafu);
         debug!(
             "Schedule request: {:?}, queue size: {}",
             request.key(),
@@ -122,7 +131,14 @@ where
         Ok(res)
     }
 
-    async fn stop(&self) -> error::Result<()> {
+    async fn stop(&self, await_termination: bool) -> error::Result<()> {
+        let state = if await_termination {
+            STATE_AWAIT_TERMINATION
+        } else {
+            STATE_STOP
+        };
+        self.state.store(state, Ordering::Relaxed);
+
         self.cancel_token.cancel();
         let handle = { self.join_handle.lock().unwrap().take() };
         if let Some(handle) = handle {
@@ -144,7 +160,7 @@ where
         let request_queue = Arc::new(RwLock::new(DedupDeque::default()));
         let cancel_token = CancellationToken::new();
         let task_notifier = Arc::new(Notify::new());
-
+        let state = Arc::new(AtomicI32::new(STATE_RUNNING));
         let handle_loop = HandlerLoop {
             task_notifier: task_notifier.clone(),
             req_queue: request_queue.clone(),
@@ -153,6 +169,7 @@ where
                 MaxInflightTaskLimiter::new(config.max_inflight_task),
             )])),
             request_handler: handler,
+            state: state.clone(),
         };
         let join_handle = common_runtime::spawn_bg(async move {
             debug!("Task handler loop spawned");
@@ -163,6 +180,7 @@ where
             request_queue,
             cancel_token,
             task_notifier,
+            state,
         }
     }
 
@@ -170,6 +188,11 @@ where
     #[inline]
     fn remaining_requests(&self) -> usize {
         self.request_queue.read().unwrap().len()
+    }
+
+    #[inline]
+    fn running(&self) -> bool {
+        self.state.load(Ordering::Relaxed) == STATE_RUNNING
     }
 }
 
@@ -179,6 +202,7 @@ pub struct HandlerLoop<R: Request, H: Handler> {
     pub task_notifier: Arc<Notify>,
     pub request_handler: H,
     pub limiter: Arc<CascadeRateLimiter<R>>,
+    pub state: Arc<AtomicI32>,
 }
 
 impl<R, H> HandlerLoop<R, H>
@@ -188,34 +212,49 @@ where
 {
     /// Runs scheduled requests dispatch loop.
     pub async fn run(&self) {
-        let task_notifier = self.task_notifier.clone();
         let limiter = self.limiter.clone();
-        loop {
+        while self.state.load(Ordering::Relaxed) == STATE_RUNNING {
             tokio::select! {
-                _ = task_notifier.notified() => {
-                    // poll requests as many as possible until rate limited, and then wait for
-                    // notification (some task's finished).
-                    debug!("Notified, queue size: {:?}", self.req_queue.read().unwrap().len());
-                    while let Some((task_key,  req)) = self.poll_task().await{
-                        if let Ok(token) = limiter.acquire_token(&req) {
-                            debug!("Executing request: {:?}", task_key);
-                            if let Err(e) = self.handle_request(req, token, self.task_notifier.clone()).await {
-                                error!(e; "Failed to submit request: {:?}", task_key);
-                            } else {
-                                info!("Submitted task: {:?}", task_key);
-                            }
-                        } else {
-                            // rate limited, put back to req queue to wait for next schedule
-                            debug!("Put back request {:?}, queue size: {}", task_key, self.req_queue.read().unwrap().len());
-                            self.put_back_req(task_key, req).await;
-                            break;
-                        }
-                    }
+                _ = self.task_notifier.notified() => {
+                    debug!("Notified, queue size: {:?}",self.req_queue.read().unwrap().len());
+                    self.poll_and_execute(&limiter).await;
                 }
                 _ = self.cancel_token.cancelled() => {
-                    info!("Task scheduler stopped.");
-                    return;
+                    info!("Task scheduler cancelled.");
+                    break;
                 }
+            }
+        }
+        // For correctness, we need to poll requests from fifo again.
+        if self.state.load(Ordering::Relaxed) == STATE_AWAIT_TERMINATION {
+            info!("Waiting for all pending tasks to finish.");
+            self.poll_and_execute(&limiter).await;
+        }
+        info!("Task scheduler stopped");
+    }
+
+    /// Polls and executes requests as many as possible until rate limited.
+    async fn poll_and_execute(&self, limiter: &Arc<CascadeRateLimiter<R>>) {
+        while let Some((task_key, req)) = self.poll_task().await {
+            if let Ok(token) = limiter.acquire_token(&req) {
+                debug!("Executing request: {:?}", task_key);
+                if let Err(e) = self
+                    .handle_request(req, token, self.task_notifier.clone())
+                    .await
+                {
+                    error!(e; "Failed to submit request: {:?}", task_key);
+                } else {
+                    info!("Submitted task: {:?}", task_key);
+                }
+            } else {
+                // rate limited, put back to req queue to wait for next schedule
+                debug!(
+                    "Put back request {:?}, queue size: {}",
+                    task_key,
+                    self.req_queue.read().unwrap().len()
+                );
+                self.put_back_req(task_key, req).await;
+                break;
             }
         }
     }
@@ -243,5 +282,284 @@ where
         self.request_handler
             .handle_request(req, token, finish_notifier)
             .await
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::atomic::AtomicI32;
+    use std::time::Duration;
+
+    use store_api::storage::RegionId;
+
+    use super::*;
+    use crate::scheduler::dedup_deque::DedupDeque;
+    use crate::scheduler::rate_limit::{
+        BoxedRateLimitToken, CascadeRateLimiter, MaxInflightTaskLimiter,
+    };
+    use crate::scheduler::{HandlerLoop, LocalScheduler, Scheduler, SchedulerConfig};
+
+    struct CountdownLatch {
+        counter: std::sync::Mutex<usize>,
+        notifies: std::sync::RwLock<Vec<Arc<Notify>>>,
+    }
+
+    impl CountdownLatch {
+        fn new(size: usize) -> Self {
+            Self {
+                counter: std::sync::Mutex::new(size),
+                notifies: std::sync::RwLock::new(vec![]),
+            }
+        }
+
+        fn countdown(&self) {
+            let mut counter = self.counter.lock().unwrap();
+            if *counter >= 1 {
+                *counter -= 1;
+                if *counter == 0 {
+                    let notifies = self.notifies.read().unwrap();
+                    for waiter in notifies.iter() {
+                        waiter.notify_one();
+                    }
+                }
+            }
+        }
+
+        async fn wait(&self) {
+            let notify = Arc::new(Notify::new());
+            {
+                let notify = notify.clone();
+                let mut notifies = self.notifies.write().unwrap();
+                notifies.push(notify);
+            }
+            notify.notified().await
+        }
+    }
+
+    #[tokio::test]
+    async fn test_schedule_handler() {
+        common_telemetry::init_default_ut_logging();
+        let queue = Arc::new(std::sync::RwLock::new(DedupDeque::default()));
+        let latch = Arc::new(CountdownLatch::new(2));
+        let latch_cloned = latch.clone();
+        let handler = Arc::new(HandlerLoop {
+            req_queue: queue.clone(),
+            cancel_token: Default::default(),
+            task_notifier: Arc::new(Default::default()),
+            request_handler: MockHandler {
+                cb: move || {
+                    latch_cloned.countdown();
+                },
+            },
+            limiter: Arc::new(CascadeRateLimiter::new(vec![Box::new(
+                MaxInflightTaskLimiter::new(3),
+            )])),
+            state: Arc::new(AtomicI32::default()),
+        });
+
+        let handler_cloned = handler.clone();
+        common_runtime::spawn_bg(async move { handler_cloned.run().await });
+
+        queue.write().unwrap().push_back(1, MockRequest::default());
+        handler.task_notifier.notify_one();
+        queue.write().unwrap().push_back(2, MockRequest::default());
+        handler.task_notifier.notify_one();
+
+        tokio::time::timeout(Duration::from_secs(1), latch.wait())
+            .await
+            .unwrap();
+    }
+
+    #[derive(Default, Debug)]
+    struct MockRequest {
+        region_id: RegionId,
+    }
+
+    struct MockHandler<F> {
+        cb: F,
+    }
+
+    #[async_trait::async_trait]
+    impl<F> Handler for MockHandler<F>
+    where
+        F: Fn() + Send + Sync,
+    {
+        type Request = MockRequest;
+
+        async fn handle_request(
+            &self,
+            _req: Self::Request,
+            token: BoxedRateLimitToken,
+            finish_notifier: Arc<Notify>,
+        ) -> error::Result<()> {
+            (self.cb)();
+            token.try_release();
+            finish_notifier.notify_one();
+            Ok(())
+        }
+    }
+
+    impl Request for MockRequest {
+        type Key = RegionId;
+
+        fn key(&self) -> Self::Key {
+            self.region_id
+        }
+    }
+
+    #[tokio::test]
+    async fn test_scheduler() {
+        let latch = Arc::new(CountdownLatch::new(2));
+        let latch_cloned = latch.clone();
+
+        let handler = MockHandler {
+            cb: move || {
+                latch_cloned.countdown();
+            },
+        };
+        let scheduler: LocalScheduler<MockRequest> = LocalScheduler::new(
+            SchedulerConfig {
+                max_inflight_task: 3,
+            },
+            handler,
+        );
+
+        scheduler.schedule(MockRequest { region_id: 1 }).unwrap();
+
+        scheduler.schedule(MockRequest { region_id: 2 }).unwrap();
+
+        tokio::time::timeout(Duration::from_secs(1), latch.wait())
+            .await
+            .unwrap();
+    }
+
+    #[tokio::test]
+    async fn test_scheduler_many() {
+        common_telemetry::init_default_ut_logging();
+        let task_size = 100;
+
+        let latch = Arc::new(CountdownLatch::new(task_size));
+        let latch_clone = latch.clone();
+
+        let handler = MockHandler {
+            cb: move || {
+                latch_clone.countdown();
+            },
+        };
+
+        let config = SchedulerConfig {
+            max_inflight_task: 3,
+        };
+        let scheduler = LocalScheduler::new(config, handler);
+
+        for i in 0..task_size {
+            scheduler
+                .schedule(MockRequest {
+                    region_id: i as RegionId,
+                })
+                .unwrap();
+        }
+
+        tokio::time::timeout(Duration::from_secs(3), latch.wait())
+            .await
+            .unwrap();
+    }
+
+    #[tokio::test]
+    async fn test_scheduler_interval() {
+        common_telemetry::init_default_ut_logging();
+        let task_size = 100;
+        let latch = Arc::new(CountdownLatch::new(task_size));
+        let latch_clone = latch.clone();
+
+        let handler = MockHandler {
+            cb: move || {
+                latch_clone.countdown();
+            },
+        };
+
+        let config = SchedulerConfig {
+            max_inflight_task: 3,
+        };
+        let scheduler = LocalScheduler::new(config, handler);
+
+        for i in 0..task_size / 2 {
+            scheduler
+                .schedule(MockRequest {
+                    region_id: i as RegionId,
+                })
+                .unwrap();
+        }
+
+        tokio::time::sleep(Duration::from_millis(100)).await;
+        for i in task_size / 2..task_size {
+            scheduler
+                .schedule(MockRequest {
+                    region_id: i as RegionId,
+                })
+                .unwrap();
+        }
+
+        tokio::time::timeout(Duration::from_secs(6), latch.wait())
+            .await
+            .unwrap();
+    }
+
+    #[tokio::test]
+    async fn test_schedule_duplicate_tasks() {
+        common_telemetry::init_default_ut_logging();
+        let handler = MockHandler { cb: || {} };
+        let config = SchedulerConfig {
+            max_inflight_task: 30,
+        };
+        let scheduler = LocalScheduler::new(config, handler);
+
+        let mut scheduled_task = 0;
+        for _ in 0..10 {
+            if scheduler.schedule(MockRequest { region_id: 1 }).unwrap() {
+                scheduled_task += 1;
+            }
+        }
+        scheduler.stop(true).await.unwrap();
+        debug!("Schedule tasks: {}", scheduled_task);
+        assert!(scheduled_task < 10);
+    }
+
+    #[tokio::test]
+    async fn test_await_termination() {
+        common_telemetry::init_default_ut_logging();
+
+        let finished = Arc::new(AtomicI32::new(0));
+        let finished_clone = finished.clone();
+        let handler = MockHandler {
+            cb: move || {
+                finished_clone.fetch_add(1, Ordering::Relaxed);
+            },
+        };
+
+        let config = SchedulerConfig {
+            max_inflight_task: 3,
+        };
+        let scheduler = Arc::new(LocalScheduler::new(config, handler));
+        let scheduler_cloned = scheduler.clone();
+        let task_scheduled = Arc::new(AtomicI32::new(0));
+        let task_scheduled_cloned = task_scheduled.clone();
+
+        common_runtime::spawn_write(async move {
+            for i in 0..10000 {
+                if let Ok(res) = scheduler_cloned.schedule(MockRequest {
+                    region_id: i as RegionId,
+                }) {
+                    if res {
+                        task_scheduled_cloned.fetch_add(1, Ordering::Relaxed);
+                    }
+                }
+            }
+        });
+
+        tokio::time::sleep(Duration::from_millis(1)).await;
+        scheduler.stop(true).await.unwrap();
+        let finished = finished.load(Ordering::Relaxed);
+        assert_eq!(finished, task_scheduled.load(Ordering::Relaxed));
     }
 }
